Sửa lại bài 001. Tôi bị ngu.

+++ 001/001.cpp
@@ -2,12 +2,12 @@
 //
 
 #include <iostream>
-#include <algorithm>
+#include <cmath>
 using namespace std;
 
 int main()
 {
-    int x1, y1, x2, y2;
+    float x1, y1, x2, y2;
     cin >> x1 >> y1 >> x2 >> y2;
     float d = sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
     cout << d;
